refactor: extract shared posting helpers into posting.py
De-messes the eventual completer sunset. review_cli.py was a mixed module:
alongside the completer main()/review(), it held post_round, post_round_review,
post_skip_stub, _round_stamp, SKIP_MARKER, and EXPECTED_FAILURES — which the
LIVE agent path (review_agent, verify_agent) imported. That transitively chained
the agent path to the completer (and to llm.py via the model-error types bundled
into EXPECTED_FAILURES).

Now:
- new posting.py holds the posting helpers + EXPECTED_FAILURES scoped to the
  transport errors they actually catch (no llm/completer dependency).
- the completer (review_cli) keeps its own broadened catch that adds the model
  errors, so completer behavior is unchanged.
- review_agent, verify_agent, verifier_cli repoint to posting.
- posting-helper tests move to test_posting.py so they survive the sunset
  (test_clis dies with review_cli); +2 net tests.

Behavior-preserving except the agent path's EXPECTED_FAILURES no longer includes
the 3 completer model-errors it never raises. The completer path (review_cli,
verifier_cli, llm) is now a clean delete once jepa-agent #7 + egolearn #5 land.
