+++ src/autoresearch/posting.py
@@ -1,0 +1,133 @@
+"""Shared GitHub posting helpers for the reviewer and verifier.
+
+Round-numbered comments, inline reviews, and skip stubs — the machinery for
+getting a judge's findings onto a PR thread. Backend-agnostic on purpose: no
+model or completer dependency, so both the completer path and the agent path
+post through here. (Extracted from review_cli so that path can be sunset as a
+clean delete once the last repo migrates.)
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+import os
+
+from autoresearch.github import GitHubClient, GitHubError
+from autoresearch.harness import redact
+
+log = logging.getLogger(__name__)
+
+# Posting/transport failures an advisory role tolerates — logged, never fatal,
+# because an advisory reviewer or verifier must not turn a target repo's CI red.
+# Programming errors (AttributeError, KeyError, TypeError) deliberately
+# propagate. Model-call errors are NOT here: they arise only inside a completer
+# call, which catches them itself — posting has nothing to do with the model.
+EXPECTED_FAILURES = (
+    GitHubError,
+    ValueError,
+    OSError,
+    json.JSONDecodeError,
+)
+
+
+def post_round(
+    client: GitHubClient, repo: str, number: int, marker: str, body: str, pr_data: dict
+) -> str:
+    """Post one NEW comment per round — numbered, stamped with the reviewed
+    head — so every round notifies and stays visible (edits do neither).
+    Shared by the advisory reviewer and the verifier; each counts rounds by
+    ITS OWN marker. The old one-thread upsert guarded against
+    synchronize-triggered spam; runs are now only PR-open or an explicit
+    label request, so volume is human-bounded.
+    """
+    stamp, round_label = _round_stamp(client, repo, number, marker, pr_data)
+    client.comment(repo, number, body.replace(marker, f"{marker}\n{stamp}", 1))
+    return round_label
+
+
+def _round_stamp(
+    client: GitHubClient, repo: str, number: int, marker: str, pr_data: dict
+) -> tuple[str, str]:
+    """(stamp line, round label): prior rounds are counted across BOTH
+    issue comments and review bodies, so switching a role between posting
+    styles never resets its numbering."""
+    head = pr_data.get("head")
+    head_sha = str(head.get("sha", ""))[:8] if isinstance(head, dict) else ""
+    # The round number is cosmetic: an EXPECTED failure counting prior
+    # rounds must never cost the round itself. Programming errors still
+    # propagate, per this module's policy.
+    try:
+        bodies = [str(c.get("body", "")) for c in client.list_comments(repo, number)]
+        bodies += [str(r.get("body", "")) for r in client.list_pr_reviews(repo, number)]
+        # STARTS WITH the marker: a quote-reply prefixes every line with
+        # "> ", so it cannot match — and this stays true for any posting
+        # identity (Actions token, GitHub App, or a self-hoster's
+        # machine-user PAT, which posts as type User)
+        prior = [b for b in bodies if b.lstrip().startswith(marker)]
+        round_label = f"**Round {len(prior) + 1}**"
+        if head_sha and any(f"reviewed head `{head_sha}`" in b for b in prior):
+            round_label += " (re-run on the same head)"
+    except EXPECTED_FAILURES as exc:
+        log.warning("could not count prior rounds: %s", exc)
+        round_label = "**New round** (prior count unavailable)"
+    return f"{round_label} — reviewed head `{head_sha or 'unknown'}`.\n\n", round_label
+
+
+def post_round_review(
+    client: GitHubClient,
+    repo: str,
+    number: int,
+    marker: str,
+    body: str,
+    inline: list[dict],
+    pr_data: dict,
+    fallback_body: str,
+) -> str:
+    """The Reviews-API sibling of post_round: body summary plus anchored
+    inline comments, event COMMENT always (the client hard-codes it). A
+    posting failure falls back to a plain issue comment carrying
+    fallback_body — the FULL single-comment rendering, because the review
+    body alone may say no more than "findings are attached" while the
+    findings live in the rejected inline payload."""
+    stamp, round_label = _round_stamp(client, repo, number, marker, pr_data)
+    try:
+        client.create_pr_review(repo, number, body.replace(marker, f"{marker}\n{stamp}", 1), inline)
+    except EXPECTED_FAILURES as exc:
+        log.warning("inline review failed (%s); falling back to a comment", exc)
+        client.comment(repo, number, fallback_body.replace(marker, f"{marker}\n{stamp}", 1))
+    return round_label
+
+
+SKIP_MARKER = "<!-- autoresearch:round-skipped -->"
+
+
+def post_skip_stub(client: GitHubClient, repo: str, number: int, role: str, exc: Exception) -> None:
+    """Silence is invisible: when the model API refuses a round (dead
+    credits, spend cap, auth), say so on the thread instead of leaving a
+    gap only the Actions tab can see. A DIFFERENT marker than a real
+    round, deliberately — a stub never counts toward round numbering and
+    never rides as follow-up wake context (both match on their own
+    markers)."""
+    # the reviewer/verifier keys are the only secrets this process holds;
+    # auth errors are exactly the class that can echo request material
+    secrets = tuple(
+        v
+        for v in (
+            os.environ.get("ANTHROPIC_REVIEWER_KEY", ""),
+            os.environ.get("ANTHROPIC_VERIFIER_KEY", ""),
+        )
+        if v
+    )
+    note = redact(str(exc), secrets)[:200]
+    try:
+        client.comment(
+            repo,
+            number,
+            f"{SKIP_MARKER}\n*The {role} round could not run — the model API "
+            f"refused the request ({type(exc).__name__}: {note}). Treat this "
+            f"as an outage, not a clean read; re-add the review label to "
+            f"re-request once the API recovers.*",
+        )
+    except EXPECTED_FAILURES as post_exc:
+        log.warning("could not post the skip stub: %s", post_exc)
